fix(admin): cast int parameter in bundle variant lookup (v0.9.130) (#138)
* fix(admin): cast int parameter in bundle variant lookup

GetVariantIDByBundleRef used `$2::text` in one branch and bare `$2` against an int column in another. lib/pq announces one type per parameter, so Postgres errored with `operator does not exist: integer = text` and the lookup returned ErrNoRows-style failure — the importer then silently dropped the bundle slot (e.g. bundle 12414 lost its InteriorDetailer component entirely under v0.9.129). Cast `$2::int` on the column comparison so both usages have explicit types.


* chore: bump version to v0.9.130


---------

+++ backend/internal/shop/product_service.go
@@ -1538,6 +1538,10 @@ func (s *ProductService) GetVariantIDByBundleRef(
 	wcVariationID int,
 ) (string, error) {
 	var id string
+	// Cast both $2 usages: lib/pq announces one type per parameter, so a
+	// mix of `$2::text` and bare `$2` against an int column would fail with
+	// `operator does not exist: integer = text`. Explicit casts on both
+	// sides let the driver choose any type.
 	err := s.db.QueryRowContext(ctx, `
 		SELECT pv.id
 		  FROM product_variants pv
@@ -1545,7 +1549,7 @@ func (s *ProductService) GetVariantIDByBundleRef(
 		 WHERE pv.product_id = $1
 		   AND (
 		       pv.sku = p.slug || '-' || $2::text
-		       OR pv.wc_variation_id = $2
+		       OR pv.wc_variation_id = $2::int
 		   )
 		 ORDER BY (pv.sku = p.slug || '-' || $2::text) DESC
 		 LIMIT 1`,
